feat:contenido de la seccion subscription
