--- src/components/EaPlayIcon/EaPlayIcon.js
@@ -1,1 +1,0 @@
-
