+++ src/components/Subscription/Subscription.scss
@@ -1,0 +1,16 @@
+.SubscriptionSection{
+    text-align: left;
+    height: 100%;
+    width: 100%;
+    background-image:url(../../assets/Imagenes/Subscription/hero-medium-console-7x2-xl.jpg.adapt.crop7x2.1920w.jpg);
+
+    .SubscriptionSection-contenido{
+        margin-left: 20%;
+    }
+}
+
+.SubscriptionSection-MoreInfo {
+    height: 50%;
+    width: 100%;
+    background-color: #6fb0d6;
+}
